This Serialize::Destroy method isn't actually needed anymore. Fixes weirdness from a few Serializable items we had on the stack. Added a comment about why operator< in Reference fails.

+++ modules/database/db_sql.cpp
@@ -185,7 +185,7 @@ class DBSQL : public Module, public Pipe
 	void OnSerializableDestruct(Serializable *obj) anope_override
 	{
 		Serialize::Type *s_type = obj->GetSerializableType();
-		if (s_type)
+		if (s_type && obj->id > 0)
 			this->RunBackground("DELETE FROM `" + this->prefix + s_type->GetName() + "` WHERE `id` = " + stringify(obj->id));
 		this->updated_items.erase(obj);
 	}
